+++ ChangeLog.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Textures are not duplicated anymore if they use different samplers resulting in equal Unity settings (saves memory on corner-case glTFs)
 - (Import) First-time imports work now, because it is ensured that the shaders are loaded correctly (#315)
 - (Import) HDRP >= 10.0: Alpha blended materials are not invisible anymore
+- (Import) URP >= 12.0: Alpha masked materials are correctly alpha tested now
 
 ## [4.5.0] - 2022-01-24
 ### Added